requirements.txt
requests
beautifulsoup4
schedule
python-dotenv

playwright
